move Repo class to transifex/utils/common.py
